Fetch identities for a principal.
Currently this fetches Sparkplug addresses from the ConfigDB. The Plan
is that these move into the Auth service as alternative identities, but
that isn't implemented yet.

Annoyingly this required updates to both Auth and ConfigDB; the Auth
service needed to allow principals to fetch their own information, and
the ConfigDB needed to allow searches for not-existing keys (otherwise
we might pick up a device record).