CDB: Return a config with etag
To do change detection properly we need a reliable version indication on
config entries. HTTP has a mechanism for this, the ETag header, so the
intention is for the ConfigDB to start returning sensible etags.

Currently these are generated by Express, so they are in no particular
format, and they will be 'soft': sometimes the etag will change when it
didn't need to. This should not cause a correctness problem, but the
intention is for the ConfigDB to store a UUID etag against every entry.