Don't bypass MQTT.js auto-reconnect
Bypassing the auto-reconnect in order to be able to wait for a new
AP-REQ has also bypassed the auto-resubscribe. Use the reconnect event
instead, which will not cause the reconnection to wait for the new
AP-REQ; this means we are racing, and sometimes we will attempt a
CONNECT with a replayed AP-REQ, which will fail.